1-Apps/1J-Opta-LMX/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opta-lmx"
version = "0.1.0"
description = "Private AI inference engine for Apple Silicon — MLX-native, OpenAI-compatible"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Matthew Byrden" },
]

dependencies = [
    "vllm-mlx>=0.2.6",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "huggingface-hub>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
    "ruff>=0.1",
    "mypy>=1.0",
    "openai>=1.0",
]
gguf = [
    "llama-cpp-python>=0.3.0",
]

[project.scripts]
opta-lmx = "opta_lmx.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
